requirements.txt
asgiref==3.2.7
Django==3.0.5
djangorestframework==3.11.0
Pillow==7.1.1
pkg-resources==0.0.0
psycopg2-binary==2.8.4
pytz==2019.3
sqlparse==0.3.1
django-ckeditor==5.9.0
django-js-asset==1.2.2
transliterate==1.10.2
pyyaml==5.3.1
uritemplate==3.0.1

